requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Markdown==3.5.1
python-slugify==8.0.1
Pygments==2.16.1
google-generativeai==0.3.2